requirements.txt
pip==23.3.2
accelerate==0.25.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bitsandbytes==0.42.0
black==23.12.1
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.21
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
crcmod==1.7
cryptography==41.0.7
cycler==0.12.1
dashscope==1.13.6
dashvector==1.0.9
dataclasses-json==0.6.3
datasets==2.16.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distro==1.8.0
docx2txt==0.8
einops==0.7.0
emoji==2.9.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.105.0
fastjsonschema==2.19.1
ffmpy==0.3.1
filelock==3.13.1
filetype==1.2.0
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.47.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
future==0.18.3
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.25.2
google-search-results==2.4.2
googleapis-common-protos==1.62.0
gradio==4.12.0
gradio_client==0.8.0
greenlet==3.0.2
grpcio==1.58.0
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
huggingface==0.0.1
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
interchange==2021.0.4
ipykernel==6.27.1
ipython==8.18.1
ipython-ngql==0.7.5
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.4
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.10
langdetect==1.0.9
langsmith==0.0.78
llama-hub==0.0.67
llama-index==0.9.32
loralib==0.1.2
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==4.0.1
modelscope==1.10.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.1
nbformat==5.9.2
nebula3-python==3.4.0
neo4j==5.16.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.8
numpy==1.26.1
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.6.1
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.10
oss2==2.18.4
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
pansi==2020.7.3
parso==0.8.3
pathspec==0.12.1
peft @ git+https://github.com/huggingface/peft@7120c9a2636f93c8db3fc4ae466e02f338ecead8
Pillow==10.1.0
pinecone-client==3.0.1
platformdirs==4.1.0
posthog==3.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
pulsar-client==3.3.0
pure-eval==0.2.2
py2neo==2021.2.4
pyaml==23.12.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.1
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
PyMySQL==1.1.0
pyparsing==3.1.1
pypdf==3.17.4
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.12.11
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
pyvis==0.3.2
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.6.0
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
safetensors==0.4.1
scipy==1.11.4
seaborn==0.13.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.27.0
streamlit==1.30.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.0
tiktoken==0.5.2
tinycss2==1.2.1
tokenize-rt==5.2.0
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.2
transformers-stream-generator==0.0.4
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.11.6
unstructured-client==0.15.1
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.25.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
widgetsnbextension==4.0.9
wikipedia==1.4.0
wrapt==1.16.0
xformer==1.0.1
xformers==0.0.23.post1
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.17.0
